finally everything is working as they should!
Don't know what happened to my precious code today and it suddenly
started behaving weirdly. Nonetheless I have finally understood what was
the issue and how it should be resolved
Takeaway: Everything that glitters is not gold. Sometimes you get a lot
of features out of the box but they might not work for your specific
case and you will have to do hell lot of efforts to make it work.
Peace!